Remove dtype in 02.py
Expected output is in float, not int, dtype

+++ Week 2/Excercises/02.py
@@ -1,6 +1,6 @@
 """Create and print a 4x4 NumPy array with value 1 on the borders and value 0 inside."""
 import numpy as np
 
-array = np.ones([4, 4], dtype=int)
+array = np.ones([4, 4])
 array[1:-1, 1:-1] = 0
 print(array)
